move root package to "ai.vespa"
