--- client/src/main/java/com/yahoo/vespa/client/dsl/IGroupOperation.java
@@ -1,10 +1,0 @@
-// Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.client.dsl;
-
-/*
- * interface for group operation
- */
-
-public interface IGroupOperation {
-
-}
